requirements.txt
Flask==3.0.0
pymongo==4.6.1
python-dotenv==1.0.0
Flask-CORS==4.0.0
flask-socketio==5.3.6
simple-websocket
